requirements.txt
amqp==5.3.1
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
black==25.12.0
blinker==1.9.0
CacheControl==0.14.4
celery==5.6.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
defusedxml==0.7.1
Django==6.0
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-yasg==1.21.12
firebase_admin==7.1.0
flake8==7.3.0
Flask==3.1.2
flower==2.0.1
geographiclib==2.1
geopandas==1.1.2
geopy==2.4.1
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-core==2.5.0
google-cloud-firestore==2.22.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.15.0
hyperframe==6.1.0
idna==3.11
inflection==0.5.1
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.2
MarkupSafe==3.0.3
mccabe==0.7.0
msgpack==1.1.2
mypy_extensions==1.1.0
numpy==2.4.1
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pathspec==1.0.3
pillow==12.1.0
platformdirs==4.5.1
prometheus_client==0.24.1
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==6.33.4
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
PyJWT==2.10.1
pyogrio==0.12.1
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
rsa==4.9.1
shapely==2.1.2
six==1.17.0
social-auth-core==4.8.3
sqlparse==0.5.5
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.4
